Support error-only unary message responses
The gRPC specification allows responses to be sent without any message
data if returning an error (non-zero status code). For unary message
responses, the current client implementation always returns an
"UNIMPLEMENTED" status code to the user if the message data is missing,
even if a non-zero status code was returned by the server. This change
addresses the issue by forwarding the non-zero status to the user if
found, falling back on the "UNIMPLEMENTED" status error if the status
code is zero.

Fixes #17.