Add client-side timing support
This adds a "timing" feature that aims to provide somewhat accurate
timing of when an HTTP request is sent for an RPC message and when the
response is received over the connection socket, primarily for
benchmarking purposes. The feature is disabled by default, so timing
code will not be enabled unless a crate explicitly requests it in its
dependencies.

In addition to the "timing" feature itself, this also updates the
Jenkins configuration to build, test, and lint the timing feature.